+++ sdk/spring/spring-cloud-azure-service/src/main/java/com/azure/spring/service/eventhubs/factory/EventProcessorClientBuilderFactory.java
@@ -24,6 +24,7 @@ import com.azure.spring.service.eventhubs.processor.BatchEventProcessingListener
 import com.azure.spring.service.eventhubs.processor.EventProcessingListener;
 import com.azure.spring.service.eventhubs.processor.RecordEventProcessingListener;
 import com.azure.spring.service.eventhubs.properties.EventHubProcessorDescriptor;
+import org.springframework.util.Assert;
 
 import java.util.Arrays;
 import java.util.List;
@@ -149,30 +150,16 @@ public class EventProcessorClientBuilderFactory extends AbstractAzureAmqpClientB
     private void configureProcessorListener(EventProcessorClientBuilder builder) {
         final EventHubProcessorDescriptor.Batch batch = this.processorProperties.getBatch();
 
-        if (isBatchMode()) {
-            if (processorListener instanceof BatchEventProcessingListener) {
-                builder.processEventBatch(((BatchEventProcessingListener) processorListener)::onEventBatch,
+        if (processorListener instanceof BatchEventProcessingListener) {
+            Assert.notNull(batch.getMaxSize(), "Batch max size must be provided");
+            builder.processEventBatch(((BatchEventProcessingListener) processorListener)::onEventBatch,
                     batch.getMaxSize(), batch.getMaxWaitTime());
-            } else {
-                throw new IllegalArgumentException("A " + BatchEventProcessingListener.class.getSimpleName()
-                    + " is required when configure batch processor.");
-            }
-        } else {
-            if (processorListener instanceof RecordEventProcessingListener) {
-                builder.processEvent(((RecordEventProcessingListener) processorListener)::onEvent);
-            } else {
-                throw new IllegalArgumentException("A " + RecordEventProcessingListener.class.getSimpleName()
-                    + " is required when configure record processor.");
-            }
+        } else if (processorListener instanceof RecordEventProcessingListener) {
+            builder.processEvent(((RecordEventProcessingListener) processorListener)::onEvent);
         }
         builder.processError(processorListener.getErrorContextConsumer());
         builder.processPartitionClose(processorListener.getCloseContextConsumer());
         builder.processPartitionInitialization(processorListener.getInitializationContextConsumer());
-    }
-
-    private boolean isBatchMode() {
-        final EventHubProcessorDescriptor.Batch batch = this.processorProperties.getBatch();
-        return batch.getMaxWaitTime() != null || batch.getMaxSize() != null;
     }
 
 }