+++ sdk/spring/spring-integration-azure-eventhubs/src/main/java/com/azure/spring/integration/eventhubs/inbound/EventHubInboundChannelAdapter.java
@@ -4,10 +4,13 @@
 package com.azure.spring.integration.eventhubs.inbound;
 
 import com.azure.messaging.eventhubs.EventData;
+import com.azure.messaging.eventhubs.models.EventBatchContext;
 import com.azure.messaging.eventhubs.models.EventContext;
 import com.azure.messaging.eventhubs.models.PartitionContext;
+import com.azure.spring.eventhubs.checkpoint.BatchCheckpointManager;
 import com.azure.spring.eventhubs.checkpoint.CheckpointManager;
 import com.azure.spring.eventhubs.core.EventHubProcessorContainer;
+import com.azure.spring.eventhubs.support.converter.EventHubBatchMessageConverter;
 import com.azure.spring.eventhubs.support.converter.EventHubMessageConverter;
 import com.azure.spring.messaging.AzureHeaders;
 import com.azure.spring.messaging.ListenerMode;
@@ -15,6 +18,7 @@ import com.azure.spring.messaging.checkpoint.AzureCheckpointer;
 import com.azure.spring.messaging.checkpoint.CheckpointConfig;
 import com.azure.spring.messaging.checkpoint.CheckpointMode;
 import com.azure.spring.messaging.checkpoint.Checkpointer;
+import com.azure.spring.service.eventhubs.processor.BatchEventProcessingListener;
 import com.azure.spring.service.eventhubs.processor.EventProcessingListener;
 import com.azure.spring.service.eventhubs.processor.RecordEventProcessingListener;
 import com.azure.spring.service.eventhubs.processor.consumer.CloseContextConsumer;
@@ -28,6 +32,7 @@ import org.springframework.messaging.MessageHeaders;
 import org.springframework.util.Assert;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -41,6 +46,8 @@ public class EventHubInboundChannelAdapter extends MessageProducerSupport {
     private final String consumerGroup;
     private final ListenerMode listenerMode;
     private final IntegrationRecordEventProcessingListener recordEventProcessor = new IntegrationRecordEventProcessingListener();
+    private final IntegrationBatchEventProcessingListener batchEventProcessor =
+        new IntegrationBatchEventProcessingListener();
     private final CheckpointConfig checkpointConfig;
     private EventProcessingListener listener;
     private CheckpointManager checkpointManager;
@@ -67,12 +74,16 @@ public class EventHubInboundChannelAdapter extends MessageProducerSupport {
 
     @Override
     protected void onInit() {
-        if (ListenerMode.RECORD.equals(this.listenerMode)) {
+        if(ListenerMode.RECORD.equals(this.listenerMode)) {
             this.listener = recordEventProcessor;
+            this.checkpointManager = CheckpointManager.of(checkpointConfig);
+        }else{
+            this.listener = batchEventProcessor;
+            this.checkpointManager  = BatchCheckpointManager.of(checkpointConfig);
         }
 
         this.processorContainer.subscribe(this.eventHubName, this.consumerGroup, this.listener);
-        this.checkpointManager = CheckpointManager.of(checkpointConfig);
+
     }
 
     @Override
@@ -151,6 +162,62 @@ public class EventHubInboundChannelAdapter extends MessageProducerSupport {
 
         public void setPayloadType(Class<?> payloadType) {
             this.payloadType = payloadType;
+        }
+    }
+
+    private class IntegrationBatchEventProcessingListener implements BatchEventProcessingListener {
+
+        private EventHubBatchMessageConverter messageConverter = new EventHubBatchMessageConverter();
+        private Class<?> payloadType = byte[].class;
+
+
+        @Override
+        public ErrorContextConsumer getErrorContextConsumer() {
+            return errorContext -> LOGGER.error("Error occurred on partition: {}. Error: {}",
+                errorContext.getPartitionContext().getPartitionId(),
+                errorContext.getThrowable());
+        }
+
+
+        @Override
+        public CloseContextConsumer getCloseContextConsumer() {
+            return closeContext -> LOGGER.info("Stopped receiving on partition: {}. Reason: {}",
+                closeContext.getPartitionContext().getPartitionId(),
+                closeContext.getCloseReason());
+        }
+
+        @Override
+        public InitializationContextConsumer getInitializationContextConsumer() {
+            return initializationContext -> LOGGER.info("Started receiving on partition: {}",
+                initializationContext.getPartitionContext().getPartitionId());
+        }
+
+        public void setMessageConverter(EventHubBatchMessageConverter converter) {
+            this.messageConverter = converter;
+        }
+
+        public void setPayloadType(Class<?> payloadType) {
+            this.payloadType = payloadType;
+        }
+
+        @Override
+        public void onEventBatch(EventBatchContext eventBatchContext) {
+            PartitionContext partition = eventBatchContext.getPartitionContext();
+
+            Map<String, Object> headers = new HashMap<>();
+            headers.put(AzureHeaders.RAW_PARTITION_ID, partition.getPartitionId());
+
+            Checkpointer checkpointer = new AzureCheckpointer(eventBatchContext::updateCheckpointAsync);
+            if (CheckpointMode.MANUAL.equals(checkpointConfig.getMode())) {
+                headers.put(AzureHeaders.CHECKPOINTER, checkpointer);
+            }
+
+            Message<?> message = this.messageConverter.toMessage(eventBatchContext, new MessageHeaders(headers), payloadType);
+
+            sendMessage(message);
+            if(checkpointConfig.getMode().equals(CheckpointMode.BATCH)) {
+                ((BatchCheckpointManager)checkpointManager).onMessage(eventBatchContext);
+            }
         }
     }