+++ sdk/spring/spring-messaging-azure-eventhubs/src/main/java/com/azure/spring/eventhubs/checkpoint/BatchCheckpointManager.java
@@ -4,6 +4,7 @@
 package com.azure.spring.eventhubs.checkpoint;
 
 import com.azure.messaging.eventhubs.EventData;
+import com.azure.messaging.eventhubs.models.EventBatchContext;
 import com.azure.messaging.eventhubs.models.EventContext;
 import com.azure.spring.messaging.checkpoint.CheckpointConfig;
 import com.azure.spring.messaging.checkpoint.CheckpointMode;
@@ -11,17 +12,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.Assert;
 
+import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Do checkpoint after each batch. Effective when {@link CheckpointMode#BATCH}
- * @deprecated suggest other mode
+ *
  * @author Warren Zhu
  * @author Xiaolu Dai
+
  */
-@Deprecated
-class BatchCheckpointManager extends CheckpointManager {
+public class BatchCheckpointManager extends CheckpointManager {
     private static final Logger LOG = LoggerFactory.getLogger(BatchCheckpointManager.class);
+    private static final String CHECKPOINT_FAIL_MSG = "Consumer group '%s' failed to checkpoint offset %s of message "
+        + "%s on partition %s, last checkpointed message is %s";
+    private static final String CHECKPOINT_SUCCESS_MSG =
+        "Consumer group '%s' checkpointed offset %s of message %s on partition %s in %s " + "mode";
     private final ConcurrentHashMap<String, EventData> lastEventByPartition = new ConcurrentHashMap<>();
 
     BatchCheckpointManager(CheckpointConfig checkpointConfig) {
@@ -30,19 +36,42 @@ class BatchCheckpointManager extends CheckpointManager {
             () -> "BatchCheckpointManager should have checkpointMode batch");
     }
 
-    @Override
-    public void onMessage(EventContext context, EventData eventData) {
-        this.lastEventByPartition.put(context.getPartitionContext().getPartitionId(), eventData);
+    public void onMessage(EventBatchContext context) {
+        EventData lastEvent = getLastEnqueuedEvent(context);
+        Long offset = lastEvent.getOffset();
+        context.updateCheckpointAsync()
+            .doOnError(t -> logCheckpointFail(context, offset, lastEvent,
+                lastEventByPartition.get(context.getPartitionContext().getPartitionId()), t))
+            .doOnSuccess(v -> {
+                this.lastEventByPartition.put(context.getPartitionContext().getPartitionId(), lastEvent);
+                logCheckpointSuccess(context, offset, lastEvent);
+            })
+            .subscribe();
     }
 
-    @Override
-    public void completeBatch(EventContext context) {
-        EventData eventData = this.lastEventByPartition.get(context.getPartitionContext().getPartitionId());
 
-        context.updateCheckpointAsync()
-            .doOnError(t -> logCheckpointFail(context, eventData, t))
-            .doOnSuccess(v -> logCheckpointSuccess(context, eventData))
-            .subscribe();
+    void logCheckpointFail(EventBatchContext context, Long offset, EventData lastEnqueuedEvent,
+                           EventData lastCheckpointedEvent, Throwable t) {
+        if (getLogger().isWarnEnabled()) {
+            getLogger().warn(String
+                .format(CHECKPOINT_FAIL_MSG, context.getPartitionContext().getConsumerGroup(), offset,
+                    lastEnqueuedEvent, lastCheckpointedEvent, context.getPartitionContext().getPartitionId()), t);
+        }
+    }
+
+    void logCheckpointSuccess(EventBatchContext context, Long offset, EventData lastEnqueuedEvent) {
+        if (getLogger().isDebugEnabled()) {
+            getLogger().debug(String
+                .format(CHECKPOINT_SUCCESS_MSG, context.getPartitionContext().getConsumerGroup(), offset,
+                    lastEnqueuedEvent, context.getPartitionContext().getPartitionId(),
+                    this.checkpointConfig.getMode()));
+        }
+    }
+
+    EventData getLastEnqueuedEvent(EventBatchContext context) {
+        List<EventData> events = context.getEvents();
+        return events.get(events.size() - 1);
+
     }
 
     @Override