+++ sdk/spring/spring-messaging-azure-eventhubs/src/main/java/com/azure/spring/eventhubs/checkpoint/CheckpointManager.java
@@ -55,7 +55,7 @@ public abstract class CheckpointManager {
         // no-op
     }
 
-    @Deprecated
+
     public void completeBatch(EventContext context) {
         // no-op
     }