+++ sdk/spring/spring-messaging-azure-eventhubs/src/main/java/com/azure/spring/eventhubs/core/processor/DefaultEventHubNamespaceProcessorFactory.java
@@ -55,6 +55,7 @@ public class DefaultEventHubNamespaceProcessorFactory implements EventHubProcess
                                                         ProcessorProperties> supplier) {
         this(checkpointStore, null, supplier);
     }
+
     public DefaultEventHubNamespaceProcessorFactory(CheckpointStore checkpointStore,
                                                     NamespaceProperties namespaceProperties,
                                                     PropertiesSupplier<Tuple2<String, String>,