+++ sdk/spring/spring-messaging-azure/src/main/java/com/azure/spring/messaging/converter/AbstractAzureMessageConverter.java
@@ -57,11 +57,11 @@ public abstract class AbstractAzureMessageConverter<I, O> implements AzureMessag
      * @return The converted object.
      * @throws ConversionException When fail to convert to object from byte array.
      */
-    private <U> U fromPayload(byte[] payload, Class<U> payloadType) {
+    protected <U> U fromPayload(Object payload, Class<U> payloadType) {
         try {
-            return getObjectMapper().readerFor(payloadType).readValue(payload);
+            return getObjectMapper().readerFor(payloadType).readValue((byte[])payload);
         } catch (IOException e) {
-            throw new ConversionException("Failed to read JSON: " + Arrays.toString(payload), e);
+            throw new ConversionException("Failed to read JSON: " + Arrays.toString((byte[])payload), e);
         }
     }
 
@@ -85,7 +85,7 @@ public abstract class AbstractAzureMessageConverter<I, O> implements AzureMessag
         return (Message<U>) internalToMessage(azureMessage, mergedHeaders, targetPayloadClass);
     }
 
-    protected abstract byte[] getPayload(I azureMessage);
+    protected abstract Object getPayload(I azureMessage);
 
     protected abstract O fromString(String payload);
 
@@ -121,15 +121,15 @@ public abstract class AbstractAzureMessageConverter<I, O> implements AzureMessag
         return fromByte(toPayload(payload));
     }
 
-    private <U> Message<?> internalToMessage(I azureMessage, Map<String, Object> headers, Class<U> targetPayloadClass) {
-        byte[] payload = getPayload(azureMessage);
-        Assert.isTrue(payload != null && payload.length > 0, "payload must not be null");
+    protected  <U> Message<?> internalToMessage(I azureMessage, Map<String, Object> headers, Class<U> targetPayloadClass) {
+        Object payload = getPayload(azureMessage);
+        Assert.isTrue(payload != null, "payload must not be null");
         if (targetPayloadClass.isInstance(azureMessage)) {
             return MessageBuilder.withPayload(azureMessage).copyHeaders(headers).build();
         }
 
         if (targetPayloadClass == String.class) {
-            return MessageBuilder.withPayload(new String(payload, StandardCharsets.UTF_8)).copyHeaders(headers).build();
+            return MessageBuilder.withPayload(new String((byte[])payload, StandardCharsets.UTF_8)).copyHeaders(headers).build();
         }
 
         if (targetPayloadClass == byte[].class) {